move huge select stories
